Update book.js file
